Use 64 bit values for flag debug printing.

+++ fuse/typeprint.go
@@ -7,8 +7,8 @@ import (
 
 func (me *WithFlags) String() string {
 	return fmt.Sprintf("File %s (%s) %s %s",
-		me.File, me.Description, raw.FlagString(raw.OpenFlagNames, int(me.OpenFlags), "O_RDONLY"),
-		raw.FlagString(raw.FuseOpenFlagNames, int(me.FuseFlags), ""))
+		me.File, me.Description, raw.FlagString(raw.OpenFlagNames, int64(me.OpenFlags), "O_RDONLY"),
+		raw.FlagString(raw.FuseOpenFlagNames, int64(me.FuseFlags), ""))
 }
 
 func (a *Attr) String() string {
