+++ raw/print_linux.go
@@ -33,15 +33,15 @@ func (me *GetAttrIn) String() string {
 func (me *ReadIn) String() string {
 	return fmt.Sprintf("{Fh %d off %d sz %d %s L %d %s}",
 		me.Fh, me.Offset, me.Size,
-		FlagString(readFlagNames, int(me.ReadFlags), ""),
+		FlagString(readFlagNames, int64(me.ReadFlags), ""),
 		me.LockOwner,
-		FlagString(OpenFlagNames, int(me.Flags), "RDONLY"))
+		FlagString(OpenFlagNames, int64(me.Flags), "RDONLY"))
 }
 
 func (me *WriteIn) String() string {
 	return fmt.Sprintf("{Fh %d off %d sz %d %s L %d %s}",
 		me.Fh, me.Offset, me.Size,
-		FlagString(writeFlagNames, int(me.WriteFlags), ""),
+		FlagString(writeFlagNames, int64(me.WriteFlags), ""),
 		me.LockOwner,
-		FlagString(OpenFlagNames, int(me.Flags), "RDONLY"))
+		FlagString(OpenFlagNames, int64(me.Flags), "RDONLY"))
 }
